+++ Makefile
@@ -72,6 +72,7 @@ SRC_FILES := main.c \
 				tokenizer/tokenizer_utils_4.c \
 				tokenizer/tokenizer_utils_5.c \
                 tokenizer/tokenizer.c \
+				parser/handle_expr.c \
                 parser/join_tokens.c \
 				parser/join_tokens_utils.c \
 				parser/parse_double_quotes_utils_1.c \
