--- no_permission.txt
@@ -1,1 +1,0 @@
-hi
