+++ src/parser/parse_double_quotes_utils_1.c
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 13:16:55 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/17 14:29:41 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/30 20:25:29 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,19 +52,6 @@ void	append_to_result(char **result, t_shell *minishell, const char *str)
 	minishell->j += str_len;
 }
 
-/*void	append_to_result(char **result, t_shell *minishell, const char *str)
-{
-	size_t	str_len;
-
-	str_len = ft_strlen(str);
-	if (minishell->j + str_len + 1 > minishell->allocated_size)
-	{
-		minishell->allocated_size = minishell->j + str_len + 1;
-		*result = ft_realloc(*result, minishell->allocated_size);
-	}
-	ft_strcpy(&(*result)[minishell->j], str);
-	minishell->j += str_len;
-}*/
 void	handle_variable_value(char **result, t_shell *minishell,
 		char *var_value)
 {
@@ -81,35 +68,3 @@ void	handle_variable_not_found(char **result, t_shell *minishell,
 	(void)var_name;
 	append_to_result(result, minishell, "");
 }
-
-/*void	handle_variable_not_found(char **result, t_shell *minishell,
-		const char *var_name)
-{
-	char	*temp;
-
-	temp = prepend_dollar(var_name);
-	append_to_result(result, minishell, temp);
-	free(temp);
-}*/
-
-/*void	handle_env_variable_expansion(char *str, char **result,
-		t_shell *minishell)
-{
-	char	var_name[256];
-	char	*var_value;
-	int		var_len;
-	char	var_name[256];
-	char	*var_value;
-	int		var_len;
-
-	var_len = 0;
-	minishell->i++;
-	while (ft_isalnum(str[minishell->i]) || str[minishell->i] == '_')
-		var_name[var_len++] = str[minishell->i++];
-	var_name[var_len] = '\0';
-	var_value = get_env_value(minishell, var_name);
-	if (var_value)
-		handle_variable_value(result, minishell, var_value);
-	else
-		handle_variable_not_found(result, minishell, var_name);
-}*/
