+++ src/parser/parse_value.c
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/26 16:54:35 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/29 21:28:25 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/30 20:12:42 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,20 +89,4 @@ void	parse_value(t_token *token_lst, t_shell *minishell)
 		process_token(curr, minishell);
 		curr = curr->next;
 	}
-}
-
-//SPECIAL
-void	handle_expr_exit_status_token(t_token *curr, t_shell *minishell)
-{
-	char	*expanded_expr;
-	char	*exit_status_str;
-
-	expanded_expr = ft_strdup("expr ");
-	exit_status_str = ft_itoa(minishell->last_return);
-	expanded_expr = ft_strjoin_free(&expanded_expr, exit_status_str);
-	expanded_expr = ft_strjoin_free(&expanded_expr, " + ");
-	expanded_expr = ft_strjoin_free(&expanded_expr, exit_status_str);
-	free(curr->token);
-	curr->token = expanded_expr;
-	free(exit_status_str);
 }
