+++ src/pipex/get_command_array.c
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/23 13:52:34 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/29 19:26:27 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/30 19:50:42 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
