+++ src/pipex/pipex_utils_3.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   pipex_utils_3.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 17:02:34 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/30 15:08:56 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/30 20:21:43 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,69 +25,30 @@ void	ft_free_tab(char **tab)
 	free(tab);
 }
 
-// OPTION_1
-/*static char	*try_exec_path(char **all_path, char **s_cmd)
-{
-	int		i;
-	char	*exec;
-	char	*path_part;
-
-	i = 0;
-	while (all_path[i])
-	{
-		path_part = ft_strjoin(all_path[i], "/");
-		exec = ft_strjoin(path_part, s_cmd[0]);
-		free(path_part);
-		if (access(exec, F_OK | X_OK) == 0)
-		{
-			return (exec);
-		}
-		free(exec);
-		i++;
-	}
-	return (NULL);
-}
-
-char	*get_path(char *cmd, t_shell *minishell)
-{
-	char	*exec;
-	char	**all_path;
-	char	**s_cmd;
-	char	*env_path;
-
-	env_path = get_env_value(minishell, "PATH");  // Use custom get_env_value
-	if (env_path == NULL)
-		return (NULL);
-	all_path = ft_split(env_path, ':');
-	free(env_path);  // Free the memory allocated by get_env_value
-	s_cmd = ft_split(cmd, ' ');
-	exec = try_exec_path(all_path, s_cmd);
-	ft_free_tab(all_path);
-	ft_free_tab(s_cmd);
-	return (exec);
-}*/
 void	ft_free_two_tabs(char **tab1, char **tab2)
 {
 	ft_free_tab(tab1);
 	ft_free_tab(tab2);
 }
 
-// OPTION_2 (PREFFERED)
-char	*get_path(char *cmd, t_shell *minishell)
+static void	initialize_path_components(char *cmd, t_shell *minishell,
+		char ***all_path, char ***s_cmd)
+{
+	char	*env_val;
+
+	env_val = get_env_value(minishell, "PATH", 1);
+	*all_path = ft_split(env_val, ':');
+	free(env_val);
+	*s_cmd = ft_split(cmd, ' ');
+}
+
+static char	*search_executable(char **all_path, char **s_cmd)
 {
 	int		i;
 	char	*exec;
 	char	*path_part;
-	char	**all_path;
-	char	**s_cmd;
-	char	*env_val;
 
-	(void)minishell;
 	i = -1;
-	env_val = get_env_value(minishell, "PATH", 1);
-	all_path = ft_split(env_val, ':');
-	free(env_val);
-	s_cmd = ft_split(cmd, ' ');
 	while (all_path[++i])
 	{
 		path_part = ft_strjoin(all_path[i], "/");
@@ -95,41 +56,26 @@ char	*get_path(char *cmd, t_shell *minishell)
 		free(path_part);
 		if (access(exec, F_OK | X_OK) == 0)
 		{
-			ft_free_two_tabs(s_cmd, all_path);
 			return (exec);
 		}
 		free(exec);
+	}
+	return (NULL);
+}
+
+char	*get_path(char *cmd, t_shell *minishell)
+{
+	char	**all_path;
+	char	**s_cmd;
+	char	*exec;
+
+	initialize_path_components(cmd, minishell, &all_path, &s_cmd);
+	exec = search_executable(all_path, s_cmd);
+	if (exec)
+	{
+		ft_free_two_tabs(s_cmd, all_path);
+		return (exec);
 	}
 	ft_free_two_tabs(all_path, s_cmd);
 	return (NULL);
 }
-
-// OPTION_3
-/*char	*get_path(char *cmd, t_shell *minishell)
-{
-	int		i;
-	char	*exec;
-	char	*path_part;
-	char	**all_path;
-	char	**s_cmd;
-
-	i = -1;
-	all_path = ft_split(get_env_value(minishell, "PATH"), ':');
-	s_cmd = ft_split(cmd, ' ');
-	while (all_path[++i])
-	{
-		path_part = ft_strjoin(all_path[i], "/");
-		exec = ft_strjoin(path_part, s_cmd[0]);
-		free(path_part);
-		if (access(exec, F_OK | X_OK) == 0)
-		{
-			ft_free_tab(s_cmd);
-			ft_free_tab(all_path);
-			return (exec);
-		}
-		free(exec);
-	}
-	ft_free_tab(all_path);
-	ft_free_tab(s_cmd);
-	return (NULL);
-}*/
