+++ src/sutils_2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   sutils_2.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/15 15:36:40 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/27 01:31:08 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/30 20:25:22 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
